requirements.txt
alembic==1.8.0
aniso8601==9.0.1
click==8.1.3
colorama==0.4.5
decorator==5.1.1
Flask==2.1.2
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.4
importlib-resources==5.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.16.0
marshmallow-enum==1.5.1
packaging==21.3
psycopg2==2.9.3
pyparsing==3.0.9
python-decouple==3.6
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.37
typing_extensions==4.2.0
validators==0.20.0
Werkzeug==2.1.2
zipp==3.8.0
